backend/requirements.txt
flask==2.0.1
werkzeug==2.0.3
psycopg2-binary==2.9.9
bcrypt==4.2.0
